closing #14

+++ lib/systems/src/SysClock.cpp
@@ -27,19 +27,19 @@ SysTick_s SysClock::tick(unsigned long sysMicros)
 
     // Recalculate trigger times
     if (tickReturn.triggers.trigger1000)
-        triggerTimes[TriggerIndices_e::TRIG_1000] += 1'000;
+        triggerTimes[static_cast<int>(TriggerIndices_e::TRIG_1000)] += 1'000;
     if (tickReturn.triggers.trigger500)
-        triggerTimes[TriggerIndices_e::TRIG_500] += 2'000;
+        triggerTimes[static_cast<int>(TriggerIndices_e::TRIG_500)] += 2'000;
     if (tickReturn.triggers.trigger100)
-        triggerTimes[TriggerIndices_e::TRIG_100] += 10'000;
+        triggerTimes[static_cast<int>(TriggerIndices_e::TRIG_100)] += 10'000;
     if (tickReturn.triggers.trigger50)
-        triggerTimes[TriggerIndices_e::TRIG_50] += 20'000;
+        triggerTimes[static_cast<int>(TriggerIndices_e::TRIG_50)] += 20'000;
     if (tickReturn.triggers.trigger10)
-        triggerTimes[TriggerIndices_e::TRIG_10] += 100'000;
+        triggerTimes[static_cast<int>(TriggerIndices_e::TRIG_10)] += 100'000;
     if (tickReturn.triggers.trigger5)
-        triggerTimes[TriggerIndices_e::TRIG_5] += 200'000;
+        triggerTimes[static_cast<int>(TriggerIndices_e::TRIG_5)] += 200'000;
     if (tickReturn.triggers.trigger1)
-        triggerTimes[TriggerIndices_e::TRIG_1] += 1'000'000;
+        triggerTimes[static_cast<int>(TriggerIndices_e::TRIG_1)] += 1'000'000;
 
     return tickReturn;
 }
